Correctly initialize the next_expected sequence number buffer.
When we receive the "synack" we did not correctly initialize the
next_expected sequence number from the other end. Fixing that though
requires that the packet validator code and the control flow upon
receiving packets be updated. Do this as well.

It doesn't eliminate the ack_is_too_old bug, but it is on the path of
eliminating it.