Fix a bug where the arrival of a FIN forced us into GOT_FIN
This is troubling, but fixed: We ignored to check if the packet we got
was really the right one in the packet order. Thus, we simply ended
with a situation where we moved to the GOT_FIN state too early and
thus threw out the packet that would satisfy the reorder buffer,
blocking everyone in the process.