FIN_SENT --> DESTROY touchup.
When we are in the FIN_SENT state and receive an ACK which ok's their
ends FIN packet, we should not DESTROY the socket, but just keep being
in the same state.

I am not sure I agree with this, but it is what the utp.cpp library is
doing, so for now I'll just blindly follow their code.