Fix a problem with RESET packets.
It turns out that when we get in RESET packets, we need to make them
trigger both the conn_id on the SEND and RECV side. This means we must
store two entries for each connection: CID and CID+1. We know that any
end will send in on either of these, so as long as we just forward the
packet to the right gen_utp_worker we are good to go.

This patch alters the dispatcher so it correctly does this. We don't
check the wrap-around case yet though. This has been postponed until
later.